+++ crates/nostr-relay-builder/src/builder.rs
@@ -4,6 +4,7 @@
 
 //! Relay Builder
 
+use std::borrow::Cow;
 use std::fmt;
 use std::net::{IpAddr, SocketAddr};
 use std::num::NonZeroUsize;
@@ -109,12 +110,72 @@ pub enum LocalRelayBuilderMode {
     PublicKey(PublicKey),
 }
 
-/// Generic plugin policy response
-pub enum PolicyResult {
-    /// Policy enforces that the event/query should be accepted
+/// Write policy result
+pub enum WritePolicyResult {
+    /// Continue processing the event without taking any action.
+    Nothing,
+    /// Stop processing the event and reply with a message with a status.
+    OK {
+        /// The message to be sent.
+        message: Cow<'static, str>,
+        /// Indicates whether the operation was successful.
+        status: bool,
+    },
+    /// Stop processing the event and send a notice message.
+    Notice(Cow<'static, str>),
+}
+
+impl WritePolicyResult {
+    /// Stops processing and send a success message. Check [MachineReadablePrefix]
+    #[inline]
+    pub fn ok_msg<S>(msg: S) -> Self
+    where
+        S: Into<Cow<'static, str>>,
+    {
+        Self::OK {
+            message: msg.into(),
+            status: true,
+        }
+    }
+
+    /// Stops processing and send a rejection message. Check [MachineReadablePrefix]
+    #[inline]
+    pub fn reject<S>(msg: S) -> Self
+    where
+        S: Into<Cow<'static, str>>,
+    {
+        Self::OK {
+            message: msg.into(),
+            status: false,
+        }
+    }
+}
+
+/// Query policy result
+pub enum QueryPolicyResult {
+    /// Accept the query
     Accept,
-    /// Policy enforces that the event/query should be rejected
-    Reject(String),
+    /// Reject the query
+    Reject {
+        /// The reject message prefix.
+        prefix: MachineReadablePrefix,
+        /// The reject message.
+        message: Cow<'static, str>,
+    },
+}
+
+impl QueryPolicyResult {
+    /// Reject the query
+    #[inline]
+    pub fn reject<S>(prefix: MachineReadablePrefix, msg: S) -> Self
+    where
+        S: Into<Cow<'static, str>>,
+    {
+        Self::Reject {
+            prefix,
+            message: msg.into(),
+        }
+    }
 }
 
 /// Custom policy for accepting events into the relay database
@@ -124,7 +185,7 @@ pub trait WritePolicy: fmt::Debug + Send + Sync {
         &'a self,
         event: &'a Event,
         addr: &'a SocketAddr,
-    ) -> BoxedFuture<'a, PolicyResult>;
+    ) -> BoxedFuture<'a, WritePolicyResult>;
 }
 
 /// Filters REQ's to the internal relay database
@@ -134,7 +195,7 @@ pub trait QueryPolicy: fmt::Debug + Send + Sync {
         &'a self,
         query: &'a Filter,
         addr: &'a SocketAddr,
-    ) -> BoxedFuture<'a, PolicyResult>;
+    ) -> BoxedFuture<'a, QueryPolicyResult>;
 }
 
 #[allow(missing_docs)]